pyproject.toml
[tool.poetry]
name = "sphinx-simpleopenapi"
version = "0.1.0"
description = ""
authors = ["Uriel Mandujano <5422638+konkolorado@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Sphinx = "^4.4.0"
sphinxcontrib-httpdomain = "^1.8.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.931"
black = "^21.12b0"
isort = "^5.10.1"
types-docutils = "^0.17.4"
types-setuptools = "^57.4.7"
types-PyYAML = "^6.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
